requirements.txt
django==5.0.9
djangorestframework==3.15.2
requests==2.32.3
googlemaps==4.10.0
drf-spectacular==0.28.0
openai==1.58.1